changed organization so Ben's new filters are in correct directory structure
So I moved all of my new filters into the correct directories.  Previously they were all bundled up so that they didn't affect the original filters.  This was causing me issues during development so I decided to simple delete the old things and put mine in the right place.  This should allow people to pull down this repo and use it immediately.  That being said I still need to test this.